Moved hands to separate namespace.
