--- src/Poker.NET.Engine/Evaluators/Naive/Hands/Base/IHand.cs
@@ -1,6 +1,0 @@
-namespace Poker.NET.Engine.Evaluators.Naive.Hands.Base;
-
-public interface IHand
-{
-
-}
