+++ src/Poker.NET.Engine/Hands/Base/IHand.cs
@@ -1,0 +1,6 @@
+namespace Poker.NET.Engine.Hands.Base;
+
+public interface IHand
+{
+
+}
