+++ src/Poker.NET.Engine/Hands/FullHouse.cs
@@ -1,7 +1,7 @@
-using Poker.NET.Engine.Evaluators.Naive.Hands.Base;
+using Poker.NET.Engine.Hands.Base;
 using Poker.NET.Engine.Helpers;
 
-namespace Poker.NET.Engine.Evaluators.Naive.Hands;
+namespace Poker.NET.Engine.Hands;
 
 public readonly struct FullHouse : IHand<FullHouse>
 {
@@ -29,7 +29,7 @@ public readonly struct FullHouse : IHand<FullHouse>
             .OrderByDescending(rs => rs.ThreeOfAKindRank)
                 .ThenByDescending(rs => rs.PairRank);
         if (!matchingFullHouseRanks.Any()) throw new ArgumentException($"The hold'em hand {cards.ToCardString()} does not contain any full houses.");
-        
+
         (Rank threeOfAKindRank, Rank pairRank) = matchingFullHouseRanks.First();
         return new(threeOfAKindRank, pairRank);
     }
