+++ src/Poker.NET.Engine/Helpers/CardHelper.cs
@@ -137,6 +137,55 @@ public static class CardHelper
         return Suits[card];
     }
     
+    public static bool AreTheSameSuit(this Cards cards)
+        => (cards & Cards.Clubs) == cards
+        || (cards & Cards.Diamonds) == cards
+        || (cards & Cards.Hearts) == cards
+        || (cards & Cards.Spades) == cards;
+
+    public static bool AreTheSameRank(this Cards cards)
+        => (cards & Cards.Twos) == cards
+        || (cards & Cards.Threes) == cards
+        || (cards & Cards.Fours) == cards
+        || (cards & Cards.Fives) == cards
+        || (cards & Cards.Sixes) == cards
+        || (cards & Cards.Sevens) == cards
+        || (cards & Cards.Eights) == cards
+        || (cards & Cards.Nines) == cards
+        || (cards & Cards.Tens) == cards
+        || (cards & Cards.Jacks) == cards
+        || (cards & Cards.Queens) == cards
+        || (cards & Cards.Kings) == cards
+        || (cards & Cards.Aces) == cards;
+    
+    public static bool ContainsRanks(this Cards cards, params Rank[] ranks)
+    {
+        foreach (Rank rank in ranks)
+        {
+            // Convert the rank enum value to its corresponding Cards value
+            Cards cardsRank = rank switch
+            {
+                Rank.Two => Cards.Twos,
+                Rank.Three => Cards.Threes,
+                Rank.Four => Cards.Fours,
+                Rank.Five => Cards.Fives,
+                Rank.Six => Cards.Sixes,
+                Rank.Seven => Cards.Sevens,
+                Rank.Eight => Cards.Eights,
+                Rank.Nine => Cards.Nines,
+                Rank.Ten => Cards.Tens,
+                Rank.Jack => Cards.Jacks,
+                Rank.Queen => Cards.Queens,
+                Rank.King => Cards.Kings,
+                Rank.AceLow or Rank.AceHigh => Cards.Aces,
+                _ => throw new ArgumentException($"Invalid rank: {rank}")
+            };
+
+            if ((cards & cardsRank) == Cards.None) return false;
+        }
+        return true;
+    }
+    
     public static void ThrowIfNotSingle(Cards card)
     {
         if (card.GetCardCount() != 1) throw new ArgumentException($"Cards value of {card} does not contain exactly one card.", nameof(card));
